refactoring: create commentutils to avoid cyclic imports
